+++ src/pages/PersonalCalendar/components/LinkStatus.tsx
@@ -1,4 +1,4 @@
-import TimeTableModal from '@/pages/Calendar/components/TimetableModal/TimetableModal';
+import TimeTableModal from '@/pages/Calendar/components/TimetableModal';
 import { Link } from 'lucide-react';
 import { useState } from 'react';
